+++ mayan/apps/tags/tests/test_wizard_steps.py
@@ -399,15 +399,7 @@ class TagStepDocumentUploadWizardTestCase(
         self._clear_events()
 
         response = self._request_document_upload_wizard_post_view()
-
-        self.assertNotContains(
-            response=response, status_code=200,
-            text=self._test_tag_list[0].label
-        )
-        self.assertNotContains(
-            response=response, status_code=200,
-            text=self._test_tag_list[1].label
-        )
+        self.assertEqual(response.status_code, 302)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)